refactor: rewrote the datatypes
