+++ mojang/minecraft/net/datatypes/__init__.py
@@ -1,19 +1,23 @@
-from .array import array_t
 from .basic import (
     bool_t,
+    bytearray_t,
+    chat_t,
     float32_t,
     float64_t,
+    identifier_t,
     int8_t,
     int16_t,
     int32_t,
     int64_t,
+    string_t,
     uint8_t,
     uint16_t,
     uint32_t,
     uint64_t,
+    uuid_t,
+    varint_t,
+    varlong_t,
 )
-from .bytearray import bytearray_t, chat_t, identifier_t, string_t, uuid_t
-from .functions import enum_t, optional_t
 from .nbt import (
     Tag,
     nbt_tags_t,
@@ -28,10 +32,8 @@ from .nbt import (
     tag_short_t,
     tag_string_t,
 )
-from .varnum import varint_t, varlong_t
 
 __all__ = [
-    "array_t",
     "bool_t",
     "float32_t",
     "float64_t",
@@ -48,8 +50,6 @@ __all__ = [
     "identifier_t",
     "string_t",
     "uuid_t",
-    "enum_t",
-    "optional_t",
     "tag_byte_t",
     "tag_bytearray_t",
     "tag_double_t",
