+++ mojang/minecraft/net/datatypes/nbt.py
@@ -13,9 +13,10 @@ from typing import (
     cast,
 )
 
-from .array import _AdvancedArray
 from .basic import (
+    _bytearray_t,
     _GenericDataType,
+    _string_t,
     float32_t,
     float64_t,
     int8_t,
@@ -25,7 +26,6 @@ from .basic import (
     uint8_t,
     uint16_t,
 )
-from .bytearray import _AdvancedByteArray
 
 DataType_T = TypeVar("DataType_T")
 
@@ -41,45 +41,17 @@ def with_tag_id(
     return cast(_TagProtocol[DataType_T], data_t)
 
 
-tag_byte_t = with_tag_id(int8_t, 1)
-tag_short_t = with_tag_id(int16_t, 2)
-tag_int_t = with_tag_id(int32_t, 3)
-tag_long_t = with_tag_id(int64_t, 4)
-tag_float_t = with_tag_id(float32_t, 5)
-tag_double_t = with_tag_id(float64_t, 6)
+tag_byte_t = with_tag_id(int8_t(), 1)
+tag_short_t = with_tag_id(int16_t(), 2)
+tag_int_t = with_tag_id(int32_t(), 3)
+tag_long_t = with_tag_id(int64_t(), 4)
+tag_float_t = with_tag_id(float32_t(), 5)
+tag_double_t = with_tag_id(float64_t(), 6)
 
-tag_bytearray_t = with_tag_id(
-    _AdvancedByteArray[bytes](
-        max_len=-1,
-        length=-1,
-        len_prefix_t=int32_t,
-        encode=lambda val: val,
-        decode=lambda val: val,
-    ),
-    7,
-)
-tag_string_t = with_tag_id(
-    _AdvancedByteArray[str](
-        max_len=-1,
-        length=-1,
-        len_prefix_t=uint16_t,
-        encode=lambda val: val.encode("utf-8"),
-        decode=lambda val: val.decode("utf-8"),
-    ),
-    8,
-)
-tag_int_array_t = with_tag_id(
-    _AdvancedArray[int](
-        item_t=int32_t, max_len=-1, length=-1, len_prefix_t=int32_t
-    ),
-    11,
-)
-tag_long_array_t = with_tag_id(
-    _AdvancedArray[int](
-        item_t=int32_t, max_len=-1, length=-1, len_prefix_t=int64_t
-    ),
-    12,
-)
+tag_bytearray_t = with_tag_id(_bytearray_t(len_prefix_t=int32_t()), 7)
+tag_string_t = with_tag_id(_string_t(len_prefix_t=uint16_t()), 8)
+tag_int_array_t = with_tag_id(int32_t().array(len_prefix_t=int32_t()), 11)
+tag_long_array_t = with_tag_id(int64_t().array(len_prefix_t=int32_t()), 12)
 
 TagType = Union[
     _TagProtocol[int],
@@ -145,22 +117,22 @@ class nbt_tags_t:
         if tag.id != 9 or tag.item_id is None:
             raise RuntimeError("Invalid list tag")
 
-        written = int8_t.write(buffer, tag.item_id)
-        written += int32_t.write(buffer, len(tag.value))
+        written = int8_t().write(buffer, tag.item_id)
+        written += int32_t().write(buffer, len(tag.value))
         return written + sum(
             [cls.write_tag_payload(buffer, tag_item) for tag_item in tag.value]
         )
 
     @classmethod
     def write(cls, buffer: BinaryIO, tag: Tag) -> int:
-        written = int8_t.write(buffer, tag.id)
+        written = int8_t().write(buffer, tag.id)
         if tag.id == Tag_End.id:
             return written
 
         if tag.name is None:
             raise RuntimeError("Unexpected None tag name")
 
-        written += uint16_t.write(buffer, len(tag.name))
+        written += uint16_t().write(buffer, len(tag.name))
         written += buffer.write(tag.name.encode("utf-8"))
 
         if tag.id == 10 and isinstance(tag.value, dict):
@@ -199,8 +171,8 @@ class nbt_tags_t:
 
     @classmethod
     def read_list_tag(cls, buffer: BinaryIO):
-        item_tag_id = int8_t.read(buffer)
-        list_length = int32_t.read(buffer)
+        item_tag_id = int8_t().read(buffer)
+        list_length = int32_t().read(buffer)
 
         return [
             Tag(
@@ -210,21 +182,22 @@ class nbt_tags_t:
                 item_id=item_tag_id,
             )
             for _ in range(list_length)
-        ]
+        ], item_tag_id
 
     @classmethod
     def read(cls, buffer: BinaryIO):
-        tag_id = uint8_t.read(buffer)
+        tag_id = uint8_t().read(buffer)
         if tag_id == Tag_End.id:
             return Tag_End
 
-        name_len = uint16_t.read(buffer)
+        name_len = uint16_t().read(buffer)
         name = buffer.read(name_len).decode("utf-8")
 
         if tag_id == 10:
             return Tag(tag_id, name, cls.read_compound_tag(buffer))
 
         if tag_id == 9:
-            return Tag(tag_id, name, cls.read_list_tag(buffer))
+            list_items, item_tag_id = cls.read_list_tag(buffer)
+            return Tag(tag_id, name, list_items, item_tag_id)
 
         return Tag(tag_id, name, cls.read_tag_payload(buffer, tag_id))
